video: Moved NullVideoBackend from core to video
