+++ video/src/lib.rs
@@ -3,5 +3,6 @@ use generational_arena::Index;
 pub mod backend;
 pub mod error;
 pub mod frame;
+pub mod null;
 
 pub type VideoStreamHandle = Index;
